+++ test/unit/specs/components/Personalization/createViewCacheManager.spec.js
